\name{sampleSEP3_test}
\alias{sampleSEP3_test}
\alias{TAIR8_genes_test}
\alias{controlSEP3_test}
\docType{data}
\title{Partial dataset of a ChIP-seq experiment}
\description{
  Partial dataset of a Solexa DNA library obtained from a ChIP-seq experiment in Arabidopsis
}

\source{
  Kaufmann et al. (2009) Target Genes of the MADS Transcription Factor SEPALLATA3: Integration of Developmental and Hormonal Pathways in the $Arabidopsis$ Flower. PLoS Biol 7:e1000090
}

\examples{
data("CSAR-dataset")

}
\keyword{datasets}
